+++ src/layout/index.jsx
@@ -1,0 +1,2 @@
+export { default as PrivateLayout } from "@layout/private";
+export { default as PublicLayout } from "@layout/public";
