--- src/pages/home/Home.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Profile from "src/components/Home/Profile/Profile";
-import TitleItemBox from "src/components/Home/TitleItems/TitleItemBox";
-
-const Home = () => {
-  return (
-    <>
-      <Profile />
-      <TitleItemBox />
-    </>
-  );
-};
-
-export default Home;
