+++ src/pages/index.jsx
@@ -1,0 +1,4 @@
+export { default as Home } from "@pages/home";
+export { default as Item } from "@pages/item";
+export { default as Login } from "@pages/login";
+export { default as SignUp } from "@pages/signup";
