+++ src/pages/signup/index.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useRef, useContext } from "react";
 import { Navigate, useLocation, useNavigate } from "react-router-dom";
 import { useMutation } from "@tanstack/react-query";
-import { PcPageWrap } from "src/layout/Main/style";
+import { PcPageWrap } from "@layout/private/style";
 import CancelCheckModal from "src/components/Modal/CancelCheckModal";
 import SignUpUser from "./Form/SignUpUser";
 import SignUpItems from "./Form/SignUpItems";
